requirements.txt
pandas
matplotlib
python-chess
